refactor: simplified tree/azure-resources-view file structure
