+++ src/documentdb/ClustersExtension.ts
@@ -48,8 +48,8 @@ import { ext } from '../extensionVariables';
 import { AzureVMDiscoveryProvider } from '../plugins/service-azure-vm/AzureVMDiscoveryProvider';
 import { AzureDiscoveryProvider } from '../plugins/service-azure/AzureDiscoveryProvider';
 import { DiscoveryService } from '../services/discoveryServices';
-import { RUBranchDataProvider } from '../tree/azure-resources-view/documentdb/mongo-ru/RUBranchDataProvider';
-import { VCoreBranchDataProvider } from '../tree/azure-resources-view/documentdb/mongo-vcore/VCoreBranchDataProvider';
+import { VCoreBranchDataProvider } from '../tree/azure-resources-view/documentdb/VCoreBranchDataProvider';
+import { RUBranchDataProvider } from '../tree/azure-resources-view/mongo-ru/RUBranchDataProvider';
 import { ConnectionsBranchDataProvider } from '../tree/connections-view/ConnectionsBranchDataProvider';
 import { DiscoveryBranchDataProvider } from '../tree/discovery-view/DiscoveryBranchDataProvider';
 import {
